Fix formatting of simplify-linq-expression
